+++ CHANGELOG.md
@@ -1,12 +1,22 @@
 # Version History
 
+## Development
+
+### Changed
+
+* Use published migrations image for integration tests.
+
+### Updated
+
+* Running tests with server image 2.0.0.
+
 ## 2.0.0
 
 ### Updated
 
- * Re-generated OpenAPI docs and code for DLM 1.3.2.
- * Updated Makefile to allow for re-generation of OpenAPI docs and code.
- * Added Keykloak and Gateway to build to enable OpenAPI generation from Makefile.
+* Re-generated OpenAPI docs and code for DLM 1.3.2.
+* Updated Makefile to allow for re-generation of OpenAPI docs and code.
+* Added Keykloak and Gateway to build to enable OpenAPI generation from Makefile.
 
 ## 1.2.4